quantum_command_center/requirements.txt
pyautogen>=0.2.0
langchain>=0.1.0
pydantic>=2.0.0
openai>=1.0.0
pandas
numpy
colorama
python-dotenv
requests
